Ccxt

The Ccxt module handles the (python) ccxt library. Wraps the ccxt api.

Ccxt.MARKETS_PATH — Constant

The path to the markets data directory.

Ccxt.ccxt — Constant

The ccxt python module reference

Ccxt.ccxt_errors — Constant

Ccxt exception names

Ccxt.ccxt_ws — Constant

The ccxt.pro (websockets) python module reference

Ccxt._ccxt_errors! — Method

Populates the ccxt_errors array with error names from the ccxt library.

_ccxt_errors!()

This function checks if the ccxt_errors array is empty. If it is, it imports the ccxt.base.errors module from the ccxt library, retrieves the directory of the module, and iterates over each error. It then checks if the first character of the error name is uppercase. If it is, the error name is added to the ccxt_errors array.

Ccxt._init — Method

Initializes the Python environment and creates the markets data directory.

_init() -> Union{Nothing, Bool}

Ccxt._issupported — Method

Check if the key k is in the dictionary has and return its boolean value.

Ccxt._multifunc — Function

Choose correct ccxt function according to what the exchange supports.

Ccxt._out_as_input — Method

A dictionary for storing function wrappers with their unique identifiers.

Ccxt.ccxt_exchange — Function

Instantiate a CCXT exchange.

ccxt_exchange(name::Symbol; ...) -> Py
ccxt_exchange(name::Symbol, params; kwargs...) -> Py

This function creates an instance of a CCXT exchange. It checks if the exchange is available in the WebSocket (ws) module, otherwise it looks in the asynchronous (async) module. If optional parameters are provided, they are passed to the exchange constructor.

Ccxt.choosefunc — Method

Chooses a function based on the provided parameters and executes it.

choosefunc(
    exc,
    suffix,
    inputs::AbstractVector;
    elkey,
    kwargs...
) -> Union{Ccxt.var"#default_func#24"{Base.Pairs{Symbol, Union{}, Nothing, @NamedTuple{}}}, Ccxt.var"#multi_func#20"{Nothing, Base.Pairs{Symbol, Union{}, Nothing, @NamedTuple{}}, <:AbstractVector{T}} where T, Ccxt.var"#single_func#21"{Nothing, Base.Pairs{Symbol, Union{}, Nothing, @NamedTuple{}}, <:AbstractVector{T}} where T}

This function selects a function based on the provided exception, suffix, and inputs. It then executes the chosen function with the provided inputs and keyword arguments. The function can handle multiple types of inputs and can execute multiple functions concurrently if necessary.

Ccxt.isccxterror — Method

Determines if a Python exception is a ccxt error.

isccxterror(err::PyException) -> Bool

Ccxt.isinitialized — Method

Checks if the ccxt object is initialized.

isinitialized() -> Bool

This function checks if the global variable ccxt is initialized by checking if it's not nothing and not null in the Python context.

Ccxt.issupported — Method

Check if the key k is supported in the exc.py.has dictionary.

Ccxt.upgrade — Method

Upgrades the ccxt library to the latest version.

upgrade() -> Py

This function upgrades the ccxt library to the latest version available. It checks the current version of the ccxt library, and if a newer version is available, it upgrades the library using pip.

See Also

Exchanges - Exchange integration and configuration

Config - Exchange integration and configuration