Pbar

PlanarCore.Pbar.last_render — Constant

Stores the timestamp of the last render in the progress bar.

PlanarCore.Pbar.min_delta — Constant

Stores the minimum time difference required between two render updates.

PlanarCore.Pbar.pbar — Constant

Holds a reference to the current progress bar or nothing if no progress bar is active.

PlanarCore.Pbar.pbar_initialized — Constant

Stores whether the progress bar has been initialized.

PlanarCore.Pbar.pbar_lock — Constant

Holds a lock to avoid flickering when updating the progress bar.

PlanarCore.Pbar.pbj — Constant

The current job being rendered.

PlanarCore.Pbar.plu — Constant

The last update timestamp.

PlanarCore.Pbar.RunningJob — Type

Represents a job that is currently running in the progress bar.

job

counter

updated_at

The RunningJob struct holds a ProgressJob, a counter, and a timestamp of when it was last updated. The job field is of type ProgressJob which represents the job that is currently running. The counter field is an integer that defaults to 1 and is used to keep track of the progress of the job. The updated_at field is a DateTime object that stores the timestamp of when the job was last updated.

PlanarCore.Pbar.clearpbar — Function

Clears the current progress bar.

clearpbar()
clearpbar(pb)

The clearpbar function stops all jobs in the current progress bar, empties the job list, and then stops the progress bar itself. It uses a lock to ensure thread safety during these operations.

PlanarCore.Pbar.complete! — Function

Complete a job.

PlanarCore.Pbar.dorender — Function

Renders the progress bar if enough time has passed since the last render.

PlanarCore.Pbar.frequency! — Method

Set the update frequency globally.

PlanarCore.Pbar.pbar! — Method

Initializes a new progress bar.

pbar!(; transient, columns, kwargs...) -> Bool

The pbar! function first clears any existing progress bar, then creates a new ProgressBar with the provided arguments. The transient argument defaults to true, and columns defaults to :default.

PlanarCore.Pbar.pbclose! — Function

Terminates the progress bar.

pbclose!()
pbclose!(pb::Term.Progress.ProgressBar)
pbclose!(pb::Term.Progress.ProgressBar, all)

The pbclose! function completes all jobs in the progress bar and then stops the progress bar itself.

PlanarCore.Pbar.pbclose! — Function

Stops the progress bar after completing the job.

PlanarCore.Pbar.startjob! — Function

Starts a new job in the progress bar.

PlanarCore.Pbar.transient! — Function

Toggles pbar transient flag

PlanarCore.Pbar.@pbar! — Macro

Instantiate a progress bar:

data: length(data) determines the bar total

unit: what unit the display

desc: description will appear over the progressbar

PlanarCore.Pbar.@pbclose! — Macro

Calls pbclose! on the global progress bar.

PlanarCore.Pbar.@pbinit! — Macro

Initializes the progress bar.

PlanarCore.Pbar.@pbstop! — Macro

Stops the progress bar.

PlanarCore.Pbar.@pbupdate! — Macro

Single update to the progressbar with the new value.

PlanarCore.Pbar.@withpbar! — Macro

Same as @pbar! but with implicit closing.

The first argument should be the collection to iterate over. Optional kw arguments:

desc: description